+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -37,6 +37,7 @@
 #include "Rule18p8Check.h"
 #include "Rule19p2Check.h"
 #include "Rule8p10Check.h"
+#include "Rule8p11Check.h"
 #include "Rule8p14Check.h"
 #include "Rule8p2Check.h"
 #include "Rule8p3Check.h"
@@ -107,6 +108,8 @@ public:
         "misrac-rule-19p2");
     CheckFactories.registerCheck<Rule8p10Check>(
         "misrac-rule-8p10");
+    CheckFactories.registerCheck<Rule8p11Check>(
+        "misrac-rule-8p11");
     CheckFactories.registerCheck<Rule8p14Check>(
         "misrac-rule-8p14");
     CheckFactories.registerCheck<Rule8p2Check>(